[BUGFIX] Prevent crash when sending attachement programmatically
When sending messages programmatically with an attachment,
Swift may not have been required by TYPO3 Core.
In that case direct access to ``\Swift_Attachment`` will f
ail to autoload the file and the application will crash.

Use conditional loading because Swift will soon be
included in Composer's autoloading mechanism, as seen in
https://git.typo3.org/Packages/TYPO3.CMS.git/commit/3f1db6151b850ec2685d0ce961f767ca3455c36d

Resolves #8